Change Dockerfile setup for pip-tools
Need to do the pip-compile as a layer in
the Dockerfile. Therefore rename our production
build requirements.txt to requirements.prod.

pip-compile then creates a requirements.txt from
requirements.in.

Result is that in any Docker build we have a
requirements.prod we can use to build MkDocs and
a requirements.txt we can compare it to to see
if pinned levels have changed.